Remove AnswerDataConverter and consolidate into JsonExporter
